laravel12funzionante

+++ app/Http/Controllers/MovieController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\MovieEditRequest;
 use App\Http\Requests\MovieRequest;
 use App\Models\Movie;
@@ -10,35 +11,36 @@ class MovieController extends Controller
 {
     public function __construct()
     {
-       
-        $this->middleware('auth')->only(['create', 'store', 'edit', 'update']);
+
+        $this->middleware('auth')->except('index');
     }
 
-   
+
     public function index()
     {
         $movies = Movie::all();
         return view('movie.index', compact('movies'));
     }
 
-    
+
     public function show(Movie $movie)
     {
         return view('movie.show', compact('movie'));
     }
 
-    
+
     public function create()
     {
         return view('movie.create');
     }
 
-    
+
     public function store(MovieRequest $request)
     {
         $data = $request->validated();
 
-        
+        $data['user_id'] = Auth::id();
+
         if ($request->hasFile('img')) {
             $data['img'] = $request->file('img')->store('images', 'public');
         } else {
@@ -51,38 +53,48 @@ class MovieController extends Controller
             ->with('successMessage', 'Hai correttamente inserito il tuo film!');
     }
 
-    
+
     public function edit(Movie $movie)
     {
-        return view('movie.edit', compact('movie'));
+        if ($movie->user_id == Auth::user()->id) {
+            return view('movie.edit', compact('movie'));
+        } else {
+            return redirect()->route('homepage')
+                ->with('errorMessage', 'Non puoi vedere questa pagina');
+        }
     }
 
-    
+
     public function update(MovieEditRequest $request, Movie $movie)
     {
-        $data = $request->validated();
+        if ($movie->user_id == Auth::id()) {
 
-        
-        if ($request->hasFile('img')) {
-            $request->validate([
-                'img' => 'image',
-            ], [
-                'img.image' => 'Il file deve essere di tipo immagine',
-            ]);
-            $data['img'] = $request->file('img')->store('images', 'public');
+            $data = $request->validated();
+
+            if ($request->hasFile('img')) {
+                $data['img'] = $request->file('img')->store('images', 'public');
+            }
+
+            $movie->update($data);
+
+            return redirect()->route('homepage')
+                ->with('successMessage', 'Hai correttamente modificato il tuo film!');
+        } else {
+
+            return redirect()->route('homepage')
+                ->with('errorMessage', 'Non puoi modificare questo film!');
         }
-
-        $movie->update($data);
-
-        return redirect()->route('homepage')
-            ->with('successMessage', 'Hai correttamente modificato il tuo film!');
     }
 
     public function destroy(Movie $movie)
     {
-        $movie->delete();
-
-        return redirect()->route('homepage')
-            ->with('successMessage', 'Hai correttamente eliminato il tuo film!');
+        if ($movie->user_id == Auth::user()->id) {
+            $movie->delete();
+            return redirect()->route('homepage')
+                ->with('successMessage', 'Hai correttamente eliminato il tuo film!');
+        } else {
+            return redirect()->route('homepage')
+                ->with('errorMessage', 'Non puoi vedere questa pagina');
+        }
     }
 }
